Move NodeTypeRegistry to node.py.
